fix errors for write_credentials_spec

+++ spec/iam/write_credentials_spec.rb
@@ -21,7 +21,10 @@ $write_creds_file = "/tmp/aws_runas_write_credentials_test_#{Process.pid}"
 def setup_fresh_creds
     FileUtils.rm_f($write_creds_file)
     FileUtils.rm_f($write_creds_file + '.lock')
-    FileUtils.cp('testdata/aws_credentials', $write_creds_file)
+    if File.exist?('testdata/aws_credentials'):
+        FileUtils.cp('testdata/aws_credentials', $write_creds_file)
+    else:
+        FileUtils.touch($write_creds_file)
     FileUtils.chmod(0600, $write_creds_file)
 end
